% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb.R, R/dataset_bb.R
\name{dataset_bb}
\alias{dataset_bb}
\title{Integrals of Squared Brownian Bridge}
\usage{
dataset_bb(NUM)

dataset_bb(NUM)
}
\arguments{
\item{NUM}{Number of simulated values generated in the dataset.}
}
\value{
A vector with length equals to NUM.

A vector with length equals to NUM.
}
\description{
Generate a dataset of independent simulated values of \eqn{\int_0^1{B^2(t)}dt}, where \eqn{B} is a standard Brownian bridge on [0,1].

Generate a dataset of independent simulated values of \eqn{\int_0^1{B^2(t)}dt}, where \eqn{B} is a standard Brownian bridge on [0,1].
}
\examples{
dataset_bb(10)
dataset_bb(10)
}
