% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_bm.R
\name{dataset_bm}
\alias{dataset_bm}
\title{Integrals of Squared Brownian Motion}
\usage{
dataset_bm(NUM)
}
\arguments{
\item{NUM}{Number of simulated values generated in the dataset.}
}
\value{
A vector with length equals to NUM.
}
\description{
Generate a dataset of independent simulated values of \eqn{\int_0^1{W^2(t)}dt}, where \eqn{W} is a standard Brownian motion on [0,1].
}
\examples{
dataset_bm(10)
}
