% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_dbb.R
\name{dataset_dbb}
\alias{dataset_dbb}
\title{Integrals of Squared De-meaned Brownian Bridge}
\usage{
dataset_dbb(NUM)
}
\arguments{
\item{NUM}{Number of simulated values generated in the dataset.}
}
\value{
A vector with length equals to NUM.
}
\description{
Generate a dataset of independent simulated values of \eqn{\int_0^1{[B(t)-\int_0^1{B(x)dx}]^2}dt}, where \eqn{B} is a standard Brownian bridge on [0,1].
}
\examples{
dataset_dbb(10)
}
