% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_sbb.R
\name{dataset_sbb}
\alias{dataset_sbb}
\title{Integrals of Squared Second-level Brownian Bridge}
\usage{
dataset_sbb(NUM)
}
\arguments{
\item{NUM}{Number of simulated values generated in the dataset.}
}
\value{
A vector with length equals to NUM.
}
\description{
Generate a dataset of independent simulated values of \eqn{\int_0^1{[W(t)+(2x-3x^2)W(1)+(-6x+6x^2)\int_0^1{W(x)dx}]^2}dt}, where \eqn{W} is a standard Brownian motion on [0,1].
}
\examples{
dataset_sbb(10)
}
