% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_kernel.R
\name{default_kernel}
\alias{default_kernel}
\title{Default Kernel Function}
\usage{
default_kernel(x)
}
\arguments{
\item{x}{The variable.}
}
\value{
The kernel value.
}
\description{
Default kernel function used in calculating the estimated long-run covariance function.
}
\examples{
default_kernel(1)
default_kernel(0)
}
